+++ ocdskit/cli/commands/indent.py
@@ -34,7 +34,7 @@ class Command(BaseCommand):
                 data = self.json_load(f)
 
             with open(path, 'w') as f:
-                json.dump(data, f, indent=self.args.indent, separators=(',', ': '))
+                json.dump(data, f, ensure_ascii=False, indent=self.args.indent, separators=(',', ': '))
                 f.write('\n')
         except json.decoder.JSONDecodeError as e:
             logger.error('{} is not valid JSON. (json.decoder.JSONDecodeError: {})'.format(path, e))