+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 0.0.3
 
 * Add helpful error message if the input is not [line-delimited JSON](https://en.wikipedia.org/wiki/JSON_streaming) data.
+* Rename base exception class from `ReportError` to `OCDSKitError`.
 
 # 0.0.2 (2018-03-14)