+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 # 0.0.3
 
-* Add helpful error message if the input is not [line-delimited JSON](https://en.wikipedia.org/wiki/JSON_streaming) data.
+* Add helpful error messages if:
+  * the input is not [line-delimited JSON](https://en.wikipedia.org/wiki/JSON_streaming) data;
+  * the input to the `indent` command is not valid JSON.
 * Rename base exception class from `ReportError` to `OCDSKitError`.
 
 # 0.0.2 (2018-03-14)